+++ src/constants/jest.ts
@@ -1,3 +1,4 @@
+/* eslint-disable */
 export const JEST = {
     TRANSFORM_IGNORE_PATTERNS: ['node_modules/(@react-native-community|react-navigation|@react-navigation/.*)'],
     SETUP_FILES: ['./jestSetupFile.js', './node_modules/react-native-gesture-handler/jestSetup.js'],