+++ src/extensions/project-extensions.ts
@@ -6,13 +6,10 @@ import { GluegunToolbox } from 'gluegun';
 import { NPM7_PREFIX, QUESTIONS } from '../constants';
 import {
     assignJsTs,
-    stringToLowerCaseNoSpace,
-    Cli,
     AngularProps,
     ReactProps,
     ReactNativeProps,
     Template,
-    IonicProps,
     Language,
     getVersionString,
 } from '../utilities';
@@ -29,7 +26,7 @@ module.exports = (toolbox: GluegunToolbox): void => {
             QUESTIONS.prettier,
         ]);
 
-        const command = `${NPM7_PREFIX} && npx -p @angular/cli@^12.0.0 ng new ${name} --directory "${name}" --style=scss`;
+        const command = `${NPM7_PREFIX} && npx -p @angular/cli@^13.2.6 ng new ${name} --directory "${name}" --style=scss`;
         const spinner = print.spin('Creating a new Angular project (this may take a few minutes)...');
         const timer = system.startTimer();
         const output = await system.run(command);
@@ -98,41 +95,13 @@ module.exports = (toolbox: GluegunToolbox): void => {
         return { name, language, lint, prettier };
     };
 
-    const createIonicProject = async (): Promise<IonicProps> => {
-        const [name, lint, prettier, template]: [string, boolean, boolean, string] = await parse([
-            QUESTIONS.name,
-            QUESTIONS.lint,
-            QUESTIONS.prettier,
-            QUESTIONS.template,
-        ]);
-
-        const command = `${NPM7_PREFIX}&& npx -p @ionic/cli ionic start ${name} blank`;
-
-        const spinner = print.spin('Creating a new Ionic project (this may take a few minutes)...');
-        const timer = system.startTimer();
-        const output = await system.run(command);
-        spinner.stop();
-
-        print.info(output);
-        print.success(`Created skeleton Ionic project in ${timer() / 1000} seconds`);
-
-        return { name, lint, prettier, template };
-    };
-
     const createReactNativeProject = async (): Promise<ReactNativeProps> => {
         // Choose a name
         const [name]: [string] = await parse([QUESTIONS.name]);
 
-        // Choose a CLI
-        let [cliTemp]: [string] = await parse([QUESTIONS.cli]);
-        cliTemp = stringToLowerCaseNoSpace(cliTemp);
-        const cli: Cli = cliTemp === 'expo' ? 'expo' : 'rnc';
-
-        // Choose aa template
+        // Choose a template
         let template = '';
-        if (cli !== 'expo') {
-            [template] = await parse([QUESTIONS.template]);
-        }
+        [template] = await parse([QUESTIONS.template]);
 
         // Choose a language
         const [languageTemp]: [string] = await parse([QUESTIONS.language]);
@@ -147,16 +116,9 @@ module.exports = (toolbox: GluegunToolbox): void => {
         const [prettier] = await parse([QUESTIONS.prettier]);
 
         // Create the basic project
-        let command: string;
-        if (cli === 'expo') {
-            command = `${NPM7_PREFIX} && npx -p expo-cli expo init --name=${name} --template=${
-                isTs ? 'expo-template-blank-typescript' : 'blank'
-            } "${name}"`;
-        } else {
-            command = `${NPM7_PREFIX} && npx -p react-native@0.64.1 react-native init ${name} ${
-                isTs ? '--template react-native-template-typescript@6.6.4' : '--template react-native@0.64.1'
-            }`;
-        }
+        const command = `${NPM7_PREFIX} && npx -p react-native@0.64.1 react-native init ${name} ${
+            isTs ? '--template react-native-template-typescript@6.6.4' : '--template react-native@0.64.1'
+        }`;
 
         const spinner = print.spin('Creating a new React Native project (this may take a few minutes)...');
         const timer = system.startTimer();
@@ -166,13 +128,12 @@ module.exports = (toolbox: GluegunToolbox): void => {
         print.info(output);
         print.success(`Created skeleton React Native project in ${timer() / 1000} seconds`);
 
-        return { name, language, lint, prettier, cli, template };
+        return { name, language, lint, prettier, template };
     };
 
     toolbox.createProject = {
         angular: createAngularProject,
         react: createReactProject,
-        ionic: createIonicProject,
         reactNative: createReactNativeProject,
     };
 };