+++ src/utilities/types.ts
@@ -1,15 +1,8 @@
 export type Language = 'ts' | 'js';
-export type Cli = 'expo' | 'rnc';
 
 export type Template = 'Blank' | 'Basic Routing' | 'Authentication';
 
 export type AngularProps = {
-    name: string;
-    lint: boolean;
-    prettier: boolean;
-    template: string;
-};
-export type IonicProps = {
     name: string;
     lint: boolean;
     prettier: boolean;
@@ -26,7 +19,6 @@ export type ReactNativeProps = {
     language: Language;
     lint: boolean;
     prettier: boolean;
-    cli: Cli;
     template: string;
 };